Fixing figlet font issue
Figlet installs fonts with capitalized font names, example `Soft.flf`.  Line 18 defines a subset of the fonts to choose from.  The font names cases however are not capitalized.  The result is on case sensitive file systems the os will not be able to find the fonts that are being referred to.