No error attempting to load catalog in empty layers directory
When a user navigates to Layers/Manage Vector Layers and the layers
catalog is missing, a "No catalog at ..." error is displayed and logged.
However, if no layers have been imported, no catalog will have been
created, and this error is inappropriate.

This change updates catalogs.js functionality for getting the local
catalog. When the layers directory is empty, we no longer give an error,
and we log a message saying that the layers directory is empty and we
are skipping getting the local catalog.