Layer upload status is better displayed
When a vector dataset file is uploaded, an "Uploaded" message is displayed as text in an awkward location in between two other components. This change displays the text differently by putting it into a "toast" alert box, a technique that is often used elsewhere in the application to display status messages. This also makes a small correction in layer documentation.