Fixed case of fonts in index.js
When smk-cli is installed onto an operating system having a case sensitive filesystem, we see "no such file" errors when running commands like "smk help". This is because the font names used in index.js do not match the case of the names of the font files.

This change updates index.js to apply the correct case to fonts so that they match the names of font files.