Vector file is imported upon upload
When a vector dataset file is uploaded, an "Uploaded" message is displayed as text in an awkward location in between two other components.

This change removes the step of having to take action to import after uploading a file. The Import button is removed, and the uploaded file is imported directly after upload.

This also makes a small correction in layer documentation.