Can set showCoverageOnHover for vector layer
This adds the ability to toggle the "showCoverageOnHover" property on a vector layer.

package.json is updated to require smk version 1.0.6 or greater as the ability to read the showCoverageOnHover property in a vector layer was added in that version.

Docs are updated to reflect UI changes as well as make text more consistent.