package-lock.json added to .gitignore
This adds the package-lock.json file to .gitignore. package-lock.json is a generated file that is updated by changes in the node module tree or package.json. It is not in source control but because it changes frequently, it appears among other deliberately changed files that are in source control.