'index' no longer output as name of executable (#79)
This addresses an issue where 'index' can be seen instead of 'smk' as the name of the executable, which is caused by differing values in the node.js 'process' object across operating systems. This sets the name of the executable to always be 'smk'.

Since we no longer need the 'isGlobal' property, we remove it as well as the 'is-exec-global.js' library.

This also updates 'A application' text as 'An application'.