NPM package updates addressing vulnerabilities (#91)
This updates dependencies on two modules and also adds an override for a transitive dependency. After these changes are applied, there is still one vulnerability from the 'xmldom' package, but no fix is yet available so we cannot address it.

This also changes a 'require' call to an 'import' for the 'node-fetch' library, whose version is updated.